+++ src/scripts/common.ts
@@ -245,5 +245,28 @@ namespace Scripts {
                 Scripts.Utils.waitWhileSomethingInJournal(['You put']);
             }
         }
+
+        static hideAll() {
+            Orion.Timer('resendTime') === -1 && Orion.SetTimer('resendTime', 10000);
+            !Orion.GetGlobal('hideAll') && Orion.SetGlobal('hideAll', '0');
+            const hidden = !!parseInt(Orion.GetGlobal('hideAll'));
+            const timer = Orion.Timer('resendTime');
+
+            if (!hidden) {
+                const mobiles = Orion.FindType('any', "any", "ground", 'mobile', 15);
+                for (const m of mobiles) {
+                    Orion.Hide(m);
+                }
+                Orion.SetGlobal('hideAll', '1');
+            }
+            else if (timer >= 10000) {
+                Scripts.Utils.resetTimer('resendTime');
+                Orion.Resend();
+                Orion.SetGlobal('hideAll', '0');
+            }
+            else {
+                Scripts.Utils.log(`jeste nemuzes dat znovu resync/resend, pockej jeste ${(10000 - timer)/1000} sekund(y)`)
+            }
+        }
     }
 }